emotion component를 리액트 컴포넌트의 바깥영역으로 이동시킵니다
