+++ src/App.jsx
@@ -21,6 +21,25 @@ import { setAccessToken } from './slice';
 
 import { loadItem } from './services/storage';
 
+const Container = styled.div({
+  margin: '0 auto',
+  width: '90%',
+});
+
+const Header = styled.header({
+  margin: '2rem 0',
+  '& h1 a': {
+    display: 'block',
+    color: '#a18cd1',
+    fontSize: '2rem',
+    textAlign: 'center',
+  },
+});
+
+const Body = styled.div({
+  padding: '1rem',
+});
+
 export default function App() {
   const dispatch = useDispatch();
 
@@ -28,25 +47,6 @@ export default function App() {
   if (accessToken) {
     dispatch(setAccessToken(accessToken));
   }
-
-  const Container = styled.div({
-    margin: '0 auto',
-    width: '90%',
-  });
-
-  const Header = styled.header({
-    margin: '2rem 0',
-    '& h1 a': {
-      display: 'block',
-      color: '#a18cd1',
-      fontSize: '2rem',
-      textAlign: 'center',
-    },
-  });
-
-  const Body = styled.div({
-    padding: '1rem',
-  });
 
   return (
     <Container>
