+++ src/HomePage.jsx
@@ -4,37 +4,37 @@ import { Link } from 'react-router-dom';
 
 import styled from '@emotion/styled';
 
-export default function HomePage() {
-  const PageTitle = styled.h2({
-    marginBottom: '1rem',
-    fontSize: '2rem',
+const PageTitle = styled.h2({
+  marginBottom: '1rem',
+  fontSize: '2rem',
+  textAlign: 'center',
+});
+
+const Nav = styled.ul({
+  margin: '0 auto',
+  width: '50%',
+});
+
+const NavItem = styled.li({
+  '& + &': {
+    marginTop: '1rem',
+  },
+  '& a': {
+    display: 'block',
+    padding: '1rem',
+    fontSize: '1.2rem',
     textAlign: 'center',
-  });
-
-  const Nav = styled.ul({
-    margin: '0 auto',
-    width: '50%',
-  });
-
-  const NavItem = styled.li({
-    '& + &': {
-      marginTop: '1rem',
+    border: '1px solid #dcdcdc',
+    borderRadius: '2em',
+    transition: '0.2s linear',
+    '&:hover': {
+      background: '#a18cd1',
+      color: '#fff',
     },
-    '& a': {
-      display: 'block',
-      padding: '1rem',
-      fontSize: '1.2rem',
-      textAlign: 'center',
-      border: '1px solid #dcdcdc',
-      borderRadius: '2em',
-      transition: '0.2s linear',
-      '&:hover': {
-        background: '#a18cd1',
-        color: '#fff',
-      },
-    },
-  });
+  },
+});
 
+export default function HomePage() {
   return (
     <div>
       <PageTitle>Home</PageTitle>
