+++ src/ReviewForm.jsx
@@ -1,8 +1,22 @@
 import React from 'react';
-import { FormWrap } from './inputStyle';
 
+import styled from '@emotion/styled';
 
 import TextField from './TextField';
+
+const FormWrap = styled.div({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  margin: '2rem 0',
+  button: {
+    marginLeft: '0.8rem',
+    padding: '0.8em',
+    color: '#fff',
+    background: '#a18cd1',
+    borderRadius: '2rem',
+  },
+});
 
 function ReviewForm({ fields, onChange, onSubmit }) {
   const { score, description } = fields;
